The findByIdCustomer and findByIdAgency methods were created in CouponRepository Coupon class, CouponService and CouponController were updated

+++ src/main/java/it/unicam/cs/ids/loyaltyplatform/service/CouponService.java
@@ -2,6 +2,7 @@ package it.unicam.cs.ids.loyaltyplatform.service;
 
 import it.unicam.cs.ids.loyaltyplatform.entity.premiumprogram.Coupon;
 import it.unicam.cs.ids.loyaltyplatform.repository.CouponRepository;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -18,24 +19,32 @@ public class CouponService {
         this.couponRepository.save(coupon);
     }
 
-    public void deleteCouponById(Integer code) {
+    public void deleteCouponByCode(Integer code) {
+        if(couponRepository.findAll().parallelStream().noneMatch(x -> x.getCode().equals(code))) {
+            throw new EntityNotFoundException("The id(" + code + ") of the Coupon to delete does not exist");
+        }
         this.couponRepository.deleteById(code);
     }
 
-    public void updateCoupon(Coupon coupon) {
-        this.couponRepository.saveAndFlush(coupon);
+    public void updateCoupon(Integer code, Coupon coupon) {
+        Coupon co = this.couponRepository.findById(code).orElseThrow(() -> new EntityNotFoundException("The id(" + code + ") of the Coupon to update does not exist"));
+        co.setPoints(coupon.getPoints());
+        co.setDiscount(coupon.getDiscount());
+        co.setIdCustomer(coupon.getIdCustomer());
+        co.setIdAgency(coupon.getIdAgency());
+        this.couponRepository.saveAndFlush(co);
     }
 
-    public Coupon getCouponById(Integer code) {
-        return this.couponRepository.findById(code).orElseThrow(NullPointerException::new);
+    public Coupon getCouponByCode(Integer code) {
+        return this.couponRepository.findById(code).orElseThrow(() -> new EntityNotFoundException("The id(" + code + ") of the Coupon to get does not exist"));
     }
 
-    public List<Coupon> getCouponsByCustomerId(Integer idCustomer) {
-        return this.couponRepository.findAll().parallelStream().filter(x -> x.getIdAgency().equals(idCustomer)).toList();
+    public List<Coupon> getCouponsByIdCustomer(Integer idCustomer) {
+        return this.couponRepository.findByIdCustomer(idCustomer);
     }
 
-    public List<Coupon> getCouponsByAgencyId(Integer idAgency) {
-        return this.couponRepository.findAll().parallelStream().filter(x -> x.getIdAgency().equals(idAgency)).toList();
+    public List<Coupon> getCouponsByIdAgency(Integer idAgency) {
+        return this.couponRepository.findByIdAgency(idAgency);
     }
 
     public List<Coupon> getAllCoupons() {
